move magnetCard to own package
