+++ src/main/java/de/mari_023/fabric/ae2wtlib/terminal/FixedWTInv.java
@@ -3,7 +3,7 @@ package de.mari_023.fabric.ae2wtlib.terminal;
 import alexiil.mc.lib.attributes.Simulation;
 import alexiil.mc.lib.attributes.item.FixedItemInv;
 import alexiil.mc.lib.attributes.item.filter.ItemFilter;
-import de.mari_023.fabric.ae2wtlib.wct.ItemMagnetCard;
+import de.mari_023.fabric.ae2wtlib.wct.magnet_card.ItemMagnetCard;
 import de.mari_023.fabric.ae2wtlib.wct.WCTContainer;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.player.PlayerInventory;
